BUG: Made it so SIWU was not called for each read
Sometimes SIWU should be called when the host asks for an packet that is
less than the length of the full packet size. Other times if the full
packet is being read then the SIWU should not be called. I added a check
to see if the USB chip was busy processing the current data, if so do not
strobe SIWU. I think strobing SIWU when not needed caused a 'ZLP' or
zero length packet to transfer. This may be the reason for thos packets
that were incorrect.